parser versions merged using macro
