+++ benches/bench.rs
@@ -32,9 +32,9 @@ mod benches {
         let mut file = std::fs::File::open("8AC1B48A77DC9D0E0AE8DDC366379FFF").unwrap();
         let mut data = Vec::new();
         file.read_to_end(&mut data).unwrap();
-        let mut schema = wasm_parser::types_2_1::Schema::parse(&data).unwrap();
+        let mut schema = wasm_parser::types_2_1::Schema::parse(&data);
         b.iter(|| {
-            schema = wasm_parser::types_2_1::Schema::parse(&data).unwrap();
+            schema = wasm_parser::types_2_1::Schema::parse(&data);
         });
 
         for p in unsafe{ schema.sub_mesh_projection.primitive_type.as_slice(schema.sub_mesh_projection.len) } {
@@ -58,9 +58,9 @@ mod benches {
         let mut file = std::fs::File::open("8AC1B48A77DC9D0E0AE8DDC366379FFF").unwrap();
         let mut data = Vec::new();
         file.read_to_end(&mut data).unwrap();
-        let schema = wasm_parser::types_2_1::Schema::parse(&data).unwrap();
+        let schema = wasm_parser::types_2_1::Schema::parse(&data);
         b.iter(|| {
-            let _ = schema.children(true, |_| true).collect::<Vec<_>>();
+            let _ = schema.children(|_| true).collect::<Vec<_>>();
         });
     }
 }
